change gpio port types to uint8_t. trying to fix crash... appears to be due to json read from file.

+++ src/model/Relay.h
@@ -22,11 +22,11 @@ public:
         _config.enabled = enabled;
     }
 
-    unsigned int getPort() const {
+    uint8_t getPort() const {
         return _config.port;
     }
 
-    void setPort(unsigned int port) {
+    void setPort(uint8_t port) {
         _config.port = port;
     }
